requirements.txt
Flask==3.1.3
flask-cors==6.0.2
google-genai==1.65.0
pinecone==8.1.0
gunicorn==25.1.0
python-dotenv==1.2.2
